+++ CodeGen/Generators/DynamicTypes.py
@@ -44,7 +44,7 @@ def DynamicTypes_inl():
 	lines.append('#include <pds/EntityReader.h>')
 	lines.append('#include <pds/DynamicTypes.h>')
 	lines.append('')
-	lines.append('#include <pds/ValueTypes.inl>')
+	lines.append('#include <pds/ValueTypes.h>')
 	lines.append('')
 	lines.append('namespace pds')
 	lines.append('{')
@@ -233,8 +233,10 @@ def DynamicTypesTests_cpp():
 	lines.extend( hlp.generate_header() )
 	lines.append('#include "Tests.h"')
 	lines.append('#include <pds/DynamicTypes.h>')
-	lines.append('#include <pds/EntityWriter.inl>')
-	lines.append('#include <pds/EntityReader.inl>')
+	lines.append('#include <pds/EntityWriter.h>')
+	lines.append('#include <pds/EntityReader.h>')
+	lines.append('#include <pds/MemoryWriteStream.h>')
+	lines.append('#include <pds/MemoryReadStream.h>')
 	lines.append('')
 	lines.append('template<class _Ty> void DynamicValueTester()')
 	lines.append('    {')
