+++ CodeGen/Generators/ValueTypes.py
@@ -11,7 +11,9 @@ def CombinedTypes_h():
 	lines.append('#pragma once')
 	lines.append('')
 	lines.append('#include "pds.h"')
+	lines.append('#include "Enums.h"') 
 	lines.append('')
+	lines.append('#include <vector>')
 	lines.append('#include <ctle/idx_vector.h>')
 	lines.append('#include <ctle/optional_idx_vector.h>')
 	lines.append('#include <ctle/optional_value.h>')
@@ -29,7 +31,7 @@ def CombinedTypes_h():
 	lines.append('    using ctle::optional_vector;')
 	lines.append('')
 	lines.append('    // template to clear a combined type')
-	lines.append('    template <class _Ty> inline void clear_combined_type( _Ty& type );')
+	lines.append('    template <class _Ty> void clear_combined_type( _Ty& type );')
 	lines.append('')
 	lines.append('    // type_information stores information on the combined types in PDS')
 	lines.append('    template <class _Ty> struct combined_type_information;')
@@ -66,11 +68,11 @@ def CombinedTypes_inl():
 	def generate_clear_combined_type( base_type_name , implementing_type , container_type , item_type , num_items_per_object , base_type_combo ):
 		lines = []
 		if( container_type == 'none' ):
-			lines.append(f'    template <> inline void clear_combined_type<{base_type_combo}>( {base_type_combo} &type ) {{ type = data_type_information<{base_type_combo}>::zero; }}' )
+			lines.append(f'    template <> void clear_combined_type<{base_type_combo}>( {base_type_combo} &type ) {{ type = data_type_information<{base_type_combo}>::zero; }}' )
 		elif( container_type == 'vector' or container_type == 'idx_vector' ):
-			lines.append(f'    template <> inline void clear_combined_type<{base_type_combo}>( {base_type_combo} &type ) {{ type.clear(); }}' )
+			lines.append(f'    template <> void clear_combined_type<{base_type_combo}>( {base_type_combo} &type ) {{ type.clear(); }}' )
 		else:
-			lines.append(f'    template <> inline void clear_combined_type<{base_type_combo}>( {base_type_combo} &type ) {{ type.reset(); }}' )
+			lines.append(f'    template <> void clear_combined_type<{base_type_combo}>( {base_type_combo} &type ) {{ type.reset(); }}' )
 		return lines
 	lines.extend( hlp.function_for_all_basetype_combos( generate_clear_combined_type ))
