--- Include/pds/SHA256.h
@@ -1,27 +1,0 @@
-// pds - Persistent data structure framework, Copyright (c) 2022 Ulrik Lindahl
-// Licensed under the MIT license https://github.com/Cooolrik/pds/blob/main/LICENSE
-
-#pragma once
-
-#include "pds.h"
-
-#include <vector>
-
-namespace pds
-{
-
-// The SHA256 functions are wrappers around picosha2, and generates hash object values
-
-namespace SHA256
-{
-
-// calculate the sha256 hash of the source data
-Status CalculateHash( hash &destHash, const u8 *srcData, size_t srcDataLength );
-Status CalculateHash( u8 destDigest[32], const u8 *srcData, size_t srcDataLength );
-
-}
-// namespace SHA256
-
-
-}
-//namespace pds
