+++ Tests/EntityReadWriteTests.cpp
@@ -3,8 +3,10 @@
 
 #include "Tests.h"
 
-#include <pds/EntityReader.inl>
-#include <pds/EntityWriter.inl>
+#include <pds/EntityWriter.h>
+#include <pds/EntityReader.h>
+#include <pds/MemoryWriteStream.h>
+#include <pds/MemoryReadStream.h>
 
 template<class T> void TestEntityWriter_TestValueType( const MemoryWriteStream &ws, EntityWriter &ew, const std::vector<std::string> &key_names )
 {
