+++ Tests/HeaderLibraries.cpp
@@ -1,0 +1,11 @@
+// pds - Persistent data structure framework, Copyright (c) 2022 Ulrik Lindahl
+// Licensed under the MIT license https://github.com/Cooolrik/pds/blob/main/LICENSE
+
+#define PDS_IMPLEMENTATION
+#define CTLE_IMPLEMENTATION
+
+// used by ctle for hash<256> calculation
+#include <picosha2.h>
+
+#include <ctle/ctle.h>
+#include <pds/pds.h>
