+++ Tests/ItemTableTests.cpp
@@ -4,8 +4,10 @@
 #include "Tests.h"
 
 #include <pds/EntityValidator.h>
-#include <pds/EntityReader.inl>
-#include <pds/EntityWriter.inl>
+#include <pds/EntityWriter.h>
+#include <pds/EntityReader.h>
+#include <pds/MemoryWriteStream.h>
+#include <pds/MemoryReadStream.h>
 
 #include <pds/ItemTable_MF.h>
