+++ Tests/SectionHierarchyReadWriteTests.cpp
@@ -3,8 +3,10 @@
 
 #include "Tests.h"
 
-#include <pds/EntityReader.inl>
-#include <pds/EntityWriter.inl>
+#include <pds/EntityWriter.h>
+#include <pds/EntityReader.h>
+#include <pds/MemoryWriteStream.h>
+#include <pds/MemoryReadStream.h>
 
 class section_array;
