+++ Tests/Tests.cpp
@@ -1,7 +1,6 @@
 // pds - Persistent data structure framework, Copyright (c) 2022 Ulrik Lindahl
 // Licensed under the MIT license https://github.com/Cooolrik/pds/blob/main/LICENSE
 
-#define PDS_IMPLEMENTATION
 #include "Tests.h"
 
 #include <ctle/log.h>
