+++ pds.cmake
@@ -12,6 +12,8 @@ set(
 	./Include/pds/DirectedGraph_MF.h
 	./Include/pds/DynamicTypes.h
 	./Include/pds/DynamicTypes.inl
+	./Include/pds/EntityManager.h
+	./Include/pds/EntityManager.inl
 	./Include/pds/EntityReader.h
 	./Include/pds/EntityReader.inl
 	./Include/pds/EntityReaderTemplates.inl
@@ -20,6 +22,7 @@ set(
 	./Include/pds/EntityWriter.inl
 	./Include/pds/EntityWriterTemplates.inl
 	./Include/pds/IndexedVector.h
+	./Include/pds/IndexedVector_MF.h
 	./Include/pds/ItemTable.h
 	./Include/pds/ItemTable_MF.h
 	./Include/pds/MemoryReadStream.h
@@ -27,13 +30,12 @@ set(
 	./Include/pds/MemoryWriteStream.h
 	./Include/pds/MemoryWriteStream.inl
 	./Include/pds/pds.h
-	./Include/pds/pds.inl
-	./Include/pds/SHA256.h
-	./Include/pds/SHA256.inl
 	./Include/pds/ValueTypes.h
 	./Include/pds/ValueTypes.inl
 	./Include/pds/Varying.h
+	./Include/pds/Varying_MF.h
 	./Include/pds/Varying.inl
+	./Include/pds/Varying_MF.inl
 
 	# compilation helper files
 	./Include/pds/_pds_macros.inl
@@ -70,14 +72,14 @@ if(BUILD_PERSISTENT_DS_TESTS)
 		systemtest
 		./Tests/SystemTest.cpp 
 		./Tests/TestPackA/TestPackA.cpp 
-
+		./Tests/HeaderLibraries.cpp 
+		
 		${pds_library_files}
 		)	
 
 	target_include_directories(	
 		systemtest 
 		PUBLIC ${PROJECT_SOURCE_DIR}/Include
-		PUBLIC ${glm_SOURCE_DIR}
 		PUBLIC ${picosha2_SOURCE_DIR}
 		PUBLIC ${ctle_SOURCE_DIR}
 		PUBLIC ${PROJECT_SOURCE_DIR}/Tests 
@@ -94,6 +96,7 @@ if(BUILD_PERSISTENT_DS_TESTS)
 		${pds_library_files}
 
 		./Tests/Tests.cpp 
+		./Tests/HeaderLibraries.cpp 
 		./Tests/DirectedGraphTests.cpp
 		./Tests/DynamicTypesTests.cpp
 		./Tests/EntityReaderRandomTests.cpp
@@ -105,7 +108,7 @@ if(BUILD_PERSISTENT_DS_TESTS)
 		./Tests/TypeTests.cpp 
 		./Tests/TestHelpers/random_vals.cpp 
 		./Tests/TestPackA/TestPackA.cpp 
-
+		
 		dependencies.cmake
 		pds.cmake
 		)
@@ -113,7 +116,6 @@ if(BUILD_PERSISTENT_DS_TESTS)
 	target_include_directories( 
 		tests 
 		PUBLIC ${PROJECT_SOURCE_DIR}/Include
-		PUBLIC ${glm_SOURCE_DIR}
 		PUBLIC ${picosha2_SOURCE_DIR}
 		PUBLIC ${ctle_SOURCE_DIR}
 		PUBLIC ${PROJECT_SOURCE_DIR}/Tests
